requirements.txt
customtkinter==5.2.1
darkdetect==0.8.0
et-xmlfile==1.1.0
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.0.7
six==1.16.0
tk==0.1.0
tzdata==2023.3
